✨ feat(tab-bar)!: implements new functionality on tab-bar
Allows the tab-bar to show system battery level, date/time, WakaTime stats for the day, and the currently playing track on most music players.

BREAKING CHANGE: A compatible music player and the playerctl executable are required for the music feature. The WakaTime CLI app and a wakatime (or equivalent) api key are required for the wakatime stats. Checks are implemented to prevent crashing but the functionality will be unavailable without the prerequisites being met.